[#180855284] Refactor file name for clarity
